python/test/requirements.txt
pytest==7.1.3

# UMA & OIDC
python-keycloak==3.7.0

# Debugger for VSC
debugpy==1.4.1